% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire data set.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance from fire station in miles}
  \item{DAMAGE}{X 1000 dollars in damace}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set concerning fires
}
\details{
This is a standard data set for the course MATH4753, installed during lab 7.
}
\keyword{datasets}
