% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Iterations}

\item{x}{Sample}

\item{fun}{Function type}

\item{alpha}{Confidence Interval}

\item{cx}{some parameter}

\item{...}{additional parameters}
}
\description{
myboot2 Bootstrap Function
}
