% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp Function that shows Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Number of Iterations}

\item{lambda}{Expected Number of Events}
}
\description{
mycltp Function that shows Poisson Distribution
}
