% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm Function that finds maximum likelihood estimates for population mean and std}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{sample}

\item{mu}{population mean}

\item{sig}{sigma, standard deviation}

\item{...}{Additional Parameters}
}
\description{
mymlnorm Function that finds maximum likelihood estimates for population mean and std
}
