% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve Curve and Probability}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{a}{Value of upper limit of probability}
}
\description{
myncurve Curve and Probability
}
