% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynegbin.R
\name{mynegbin}
\alias{mynegbin}
\title{Negative Binomial function}
\usage{
mynegbin(y, r, p)
}
\arguments{
\item{y}{A random numeric value}

\item{r}{Number of elements}

\item{p}{Probability}
}
\value{
probability using negative binomial
}
\description{
Negative Binomial function
}
