% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets Number of tickets to be sold based on probability plane is overbooked}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seat in flight}

\item{gamma}{probability that plane is overbooked}

\item{p}{probabiility of a "show"}
}
\value{
A list with nd, nc, N, gamma, and p
}
\description{
ntickets Number of tickets to be sold based on probability plane is overbooked
}
