% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{quadfunc}
\alias{quadfunc}
\title{Quadratic function}
\usage{
quadfunc(x, coef, xk)
}
\arguments{
\item{x}{A quantitative vector}

\item{coef}{coefficients}

\item{xk}{A numeric value}
}
\value{
An intercept value
}
\description{
Quadratic function
}
